requirements.txt
requests>=2.31.0,<3.0.0
python-telegram-bot>=21.0,<22.0.0
python-dotenv>=1.0.0,<2.0.0
